renamed sector feature sub-folders
