+++ lib/src/config/mock/fake_sectors_status.dart
@@ -1,4 +1,4 @@
-import 'package:irrigazione_iot/src/features/sectors/domain/sector_status.dart';
+import 'package:irrigazione_iot/src/features/sectors/model/sector_status.dart';
 
 List<SectorStatus> _generateSectorStatus() {
   List<SectorStatus> sectorStatusList = [];
