+++ lib/src/config/routes/app_router.dart
@@ -10,12 +10,12 @@ import 'package:irrigazione_iot/src/features/home/screen/home_nested_navigator.d
 import 'package:irrigazione_iot/src/features/pumps/screen/add_pump/add_update_pump_form.dart';
 import 'package:irrigazione_iot/src/features/pumps/screen/pump_details/pump_details_screen.dart';
 import 'package:irrigazione_iot/src/features/pumps/screen/pump_list/pumps_list_screen.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/add_update_sector_form.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/select_a_specie_screen.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/select_an_irrigation_source.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/select_an_irrigation_system.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_details/sector_details.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_list/sectors_list_screen.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/add_update_sector_form.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/select_a_specie_screen.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/select_an_irrigation_source.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/select_an_irrigation_system.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_details/sector_details.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_list/sectors_list_screen.dart';
 import 'package:irrigazione_iot/src/features/user_companies/data/selected_company_repository.dart';
 import 'package:irrigazione_iot/src/features/user_companies/presentation/user_company_list/user_companies_list_screen.dart';
 import 'package:irrigazione_iot/src/utils/extensions.dart';
