+++ lib/src/features/sectors/data/sector_status_repository.dart
@@ -1,5 +1,5 @@
 import 'package:irrigazione_iot/src/features/sectors/data/fake_sector_status_repository.dart';
-import 'package:irrigazione_iot/src/features/sectors/domain/sector.dart';
+import 'package:irrigazione_iot/src/features/sectors/model/sector.dart';
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
