+++ lib/src/features/sectors/model/sector.dart
