+++ lib/src/features/sectors/model/specie.dart
