+++ lib/src/features/sectors/screen/add_update_sector/add_update_sector_form.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:irrigazione_iot/src/config/enums/form_types.dart';
-import 'package:irrigazione_iot/src/features/sectors/domain/sector.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/add_update_sector_controller.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/add_update_sector_form_contents.dart';
+import 'package:irrigazione_iot/src/features/sectors/model/sector.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/add_update_sector_controller.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/add_update_sector_form_contents.dart';
 import 'package:irrigazione_iot/src/utils/async_value_ui.dart';
 
 class AddUpdateSectorForm extends ConsumerWidget {
