+++ lib/src/features/sectors/screen/add_update_sector/select_a_specie_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:irrigazione_iot/src/features/sectors/data/specie_repository.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/responsive_select_screens_tile.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/responsive_select_screens_tile.dart';
 import 'package:irrigazione_iot/src/utils/extensions.dart';
 import 'package:irrigazione_iot/src/widgets/app_sliver_bar.dart';
 import 'package:irrigazione_iot/src/widgets/async_value_widget.dart';
