+++ lib/src/features/sectors/screen/add_update_sector/select_an_irrigation_source.dart
@@ -1,34 +1,33 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:irrigazione_iot/src/config/enums/irrigation_enums.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/add_update_sector/responsive_select_screens_tile.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/add_update_sector/responsive_select_screens_tile.dart';
 import 'package:irrigazione_iot/src/utils/extensions.dart';
 import 'package:irrigazione_iot/src/widgets/app_sliver_bar.dart';
 
-class SelectAnIrrigationSystem extends ConsumerWidget {
-  const SelectAnIrrigationSystem({super.key});
+class SelectAnIrrigationSource extends StatelessWidget {
+  const SelectAnIrrigationSource({super.key});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
+  Widget build(BuildContext context) {
     return Scaffold(
       body: SafeArea(
         child: CustomScrollView(
           slivers: [
             AppSliverBar(
-              title: context.loc.selectAnIrrigationSystem,
+              title: context.loc.selectAnIrrigationSource,
             ),
             SliverList(
                 delegate: SliverChildBuilderDelegate(
               (context, index) {
-                final irrigationSystemType = IrrigationSystemType.values[index];
+                final irrigationSource = IrrigationSource.values[index];
                 return ResponsiveSelectScreensTile(
-                  title: irrigationSystemType.uiName,
+                  title: irrigationSource.uiName,
                   onTap: () {
-                    Navigator.of(context).pop(irrigationSystemType.uiName);
+                    Navigator.of(context).pop(irrigationSource.uiName);
                   },
                 );
               },
-              childCount: IrrigationSystemType.values.length,
+              childCount: IrrigationSource.values.length,
             ))
           ],
         ),
