+++ lib/src/features/sectors/screen/sector_list/sector_list_tile.dart
@@ -4,11 +4,11 @@ import 'package:go_router/go_router.dart';
 import 'package:irrigazione_iot/src/config/routes/app_router.dart';
 import 'package:irrigazione_iot/src/constants/app_sizes.dart';
 import 'package:irrigazione_iot/src/constants/breakpoints.dart';
-import 'package:irrigazione_iot/src/features/sectors/domain/sector.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_list/dismiss_sector_controller.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_list/sector_list_tile_subtitle.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_switch.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_switch_controller.dart';
+import 'package:irrigazione_iot/src/features/sectors/model/sector.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_list/dismiss_sector_controller.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_list/sector_list_tile_subtitle.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_switch.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_switch_controller.dart';
 import 'package:irrigazione_iot/src/utils/custom_controller_state.dart';
 import 'package:irrigazione_iot/src/utils/extensions.dart';
 import 'package:irrigazione_iot/src/widgets/alert_dialogs.dart';
