+++ lib/src/features/sectors/screen/sector_list/sectors_list_screen.dart
@@ -5,11 +5,11 @@ import 'package:go_router/go_router.dart';
 import 'package:irrigazione_iot/src/config/enums/roles.dart';
 import 'package:irrigazione_iot/src/config/routes/app_router.dart';
 import 'package:irrigazione_iot/src/features/sectors/data/sector_repository.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/empty_sector_widget.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_list/dismiss_sector_controller.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_list/sector_list_tile.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_switch_controller.dart';
-import 'package:irrigazione_iot/src/features/sectors/presentation/sector_list/sectors_list_tile_skeleton.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/empty_sector_widget.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_list/dismiss_sector_controller.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_list/sector_list_tile.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_switch_controller.dart';
+import 'package:irrigazione_iot/src/features/sectors/screen/sector_list/sectors_list_tile_skeleton.dart';
 import 'package:irrigazione_iot/src/features/user_companies/data/user_companies_repository.dart';
 import 'package:irrigazione_iot/src/utils/async_value_ui.dart';
 import 'package:irrigazione_iot/src/utils/extensions.dart';
