+++ lib/src/features/sectors/screen/sector_list/sectors_list_tile_skeleton.dart
